// src/enums/pets.ts
export enum Species {
    Dog = 'dog',
    Cat = 'cat',
    Bird = 'bird',
    Rabbit = 'rabbit',
    Fish = 'fish',
    Rodent = 'rodent',
}

export enum Age {
    Puppy = 'puppy',
    Adult = 'adult',
    Senior = 'senior',
}

export enum Size {
    Small = 'small',
    Medium = 'medium',
    Large = 'large',
    Giant = 'giant',
}

export enum EnergyLevel {
    Low = 'low',
    Medium = 'medium',
    High = 'high',
}